# openapi.json
{
  "components": {
    "schemas": {
      "Body_upload_song_api_songs_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "visibility": {
            "$ref": "#/components/schemas/SongVisibility",
            "default": "private"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_song_api_songs_post",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "IngestJobPublic": {
        "description": "Sun Oct 4 (N4): closes the gap PROGRESS.md's Day 1 loose ends\nflagged \u2014 \"no HTTP endpoint exists yet for polling an ingest job's\nstatus,\" with tests/test_separation.py reaching into the DB directly\nas a stand-in. The SPA's progress poll (elums/api/routers/songs.py's\n`GET /songs/{song_id}/ingest`) is the real consumer.",
        "properties": {
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "ctc_alignment_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "current_stage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/IngestJobStage"
              },
              {
                "type": "null"
              }
            ]
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "f0_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "lyrics_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "note_grid_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "rms_vad_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "separation_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          },
          "step_index": {
            "title": "Step Index",
            "type": "integer"
          },
          "step_total": {
            "title": "Step Total",
            "type": "integer"
          },
          "structure_beats_status": {
            "$ref": "#/components/schemas/IngestJobStatus"
          }
        },
        "required": [
          "status",
          "current_stage",
          "separation_status",
          "structure_beats_status",
          "rms_vad_status",
          "lyrics_status",
          "ctc_alignment_status",
          "f0_status",
          "note_grid_status",
          "step_index",
          "step_total",
          "message",
          "error_message",
          "completed_at"
        ],
        "title": "IngestJobPublic",
        "type": "object"
      },
      "IngestJobStage": {
        "enum": [
          "separation",
          "structure_beats",
          "rms_vad",
          "lyrics",
          "ctc_alignment",
          "f0",
          "note_grid"
        ],
        "title": "IngestJobStage",
        "type": "string"
      },
      "IngestJobStatus": {
        "enum": [
          "pending",
          "running",
          "succeeded",
          "failed",
          "skipped"
        ],
        "title": "IngestJobStatus",
        "type": "string"
      },
...
      "PerformanceKind": {
        "enum": [
          "solo",
          "seed",
          "join"
        ],
        "title": "PerformanceKind",
        "type": "string"
      },
      "PerformancePublic": {
        "properties": {
          "alignment_warning": {
            "title": "Alignment Warning",
            "type": "boolean"
          },
          "analysis_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysis Blob Sha256"
          },
          "arrival_offset_ms_median": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Arrival Offset Ms Median"
          },
          "audio_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Blob Sha256"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "device_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Device Label"
          },
          "error_message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error Message"
          },
          "f0_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "F0 Blob Sha256"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/PerformanceKind"
          },
          "latency_offset_ms": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latency Offset Ms"
          },
          "median_cents": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Median Cents"
          },
          "octave_shift_semitones": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Octave Shift Semitones"
          },
          "offset_s": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Offset S"
          },
          "parent_performance_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Performance Id"
          },
          "pct_in_tune": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pct In Tune"
          },
          "score_overall": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score Overall"
          },
          "song_id": {
            "format": "uuid",
            "title": "Song Id",
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/PerformanceStatus"
          },
          "user_id": {
            "format": "uuid",
            "title": "User Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "song_id",
          "user_id",
          "kind",
          "parent_performance_id",
          "status",
          "error_message",
          "audio_blob_sha256",
          "f0_blob_sha256",
          "analysis_blob_sha256",
          "score_overall",
          "pct_in_tune",
          "median_cents",
          "arrival_offset_ms_median",
          "octave_shift_semitones",
          "alignment_warning",
          "offset_s",
          "latency_offset_ms",
          "device_label",
          "created_at"
        ],
        "title": "PerformancePublic",
        "type": "object"
      },
      "PerformanceStatus": {
        "enum": [
          "uploading",
          "scoring",
          "succeeded",
          "failed"
        ],
        "title": "PerformanceStatus",
        "type": "string"
      },
...
      "SongBundlePublic": {
        "description": "Tue Oct 6 (T3): `GET /api/songs/{song_id}` \u2014 song metadata plus\nthe chart/peaks/stem blob hashes the karaoke page needs. Deliberately\nNOT the chart's own contents inline \u2014 the client fetches that\nseparately through `/blobs/<chart_blob_sha256>` (Caddy's Range-request\npath), same as every other blob.",
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "chart_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Chart Blob Sha256"
          },
          "f0_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "F0 Blob Sha256"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "instrumental_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Instrumental Blob Sha256"
          },
          "note_count": {
            "title": "Note Count",
            "type": "integer"
          },
          "peaks_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Peaks Blob Sha256"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "visibility": {
            "$ref": "#/components/schemas/SongVisibility"
          },
          "vocals_blob_sha256": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Vocals Blob Sha256"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "visibility",
          "vocals_blob_sha256",
          "instrumental_blob_sha256",
          "chart_blob_sha256",
          "peaks_blob_sha256",
          "f0_blob_sha256",
          "note_count"
        ],
        "title": "SongBundlePublic",
        "type": "object"
      },
      "SongPublic": {
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "source_blob_sha256": {
            "title": "Source Blob Sha256",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "uploaded_by_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Uploaded By User Id"
          },
          "visibility": {
            "$ref": "#/components/schemas/SongVisibility"
          }
        },
        "required": [
          "id",
          "title",
          "artist",
          "source_blob_sha256",
          "visibility",
          "uploaded_by_user_id"
        ],
        "title": "SongPublic",
        "type": "object"
      },
      "SongVisibility": {
        "enum": [
          "private",
          "public"
        ],
        "title": "SongVisibility",
        "type": "string"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "Elums API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/songs": {
      "post": {
        "operationId": "upload_song_api_songs_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_song_api_songs_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Upload Song",
        "tags": [
          "songs"
        ]
      }
    },
    "/api/songs/{song_id}": {
      "get": {
        "description": "Tue Oct 6 (T3 of IMPLEMENTATION_PLAN_2026-10-06.md): the karaoke\npage's one fetch \u2014 song metadata plus the chart/peaks/stem blob\nhashes. **Reuses `get_ingest_status`'s owner-or-public rule and\n404-not-403 convention verbatim** (Day 2 \u00a75.3's precedent) rather\nthan reinventing it, per the plan's own instruction.",
        "operationId": "get_song_bundle_api_songs__song_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "song_id",
            "required": true,
            "schema": {
              "title": "Song Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SongBundlePublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Song Bundle",
        "tags": [
          "songs"
        ]
      }
    },
    "/api/songs/{song_id}/ingest": {
      "get": {
        "description": "Sun Oct 4 (N4): closes PROGRESS.md's Day 1 loose end \u2014 the SPA's\nprogress poll had no real endpoint to call and\ntests/test_separation.py stood in with a raw DB query.\n\nOwner-or-public, same rule `/internal/blob-authz` already enforces\nfor blob bytes (elums/api/routers/internal.py) \u2014 deliberately NOT\n`Depends(get_current_user)`, which hard-401s with no cookie at all;\na public song's ingest status is meant to be visible anonymously,\nthe same as its blobs.",
        "operationId": "get_ingest_status_api_songs__song_id__ingest_get",
        "parameters": [
          {
            "in": "path",
            "name": "song_id",
            "required": true,
            "schema": {
              "title": "Song Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IngestJobPublic"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Ingest Status",
        "tags": [
          "songs"
        ]
      }
    },
    "/api/songs/{song_id}/seeds": {
      "get": {
        "description": "Wed Oct 7 (W6): \u00a77.3's async seed/join \u2014 \"a session becomes a\nsolo take published as a joinable seed.\" Public by construction\n(publishing a seed is an explicit opt-in act, unlike a song's own\nvisibility default), so this endpoint takes no auth at all, same\nreasoning a public song's blobs are anonymously fetchable.",
        "operationId": "list_seeds_api_songs__song_id__seeds_get",
        "parameters": [
          {
            "in": "path",
            "name": "song_id",
            "required": true,
            "schema": {
              "title": "Song Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/PerformancePublic"
                  },
                  "title": "Response List Seeds Api Songs  Song Id  Seeds Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Seeds",
        "tags": [
          "songs"
        ]
      }
    },
...
  }
}